% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{add_check}
\alias{add_check}
\title{Add check into group}
\usage{
add_check(group, new_check)
}
\arguments{
\item{group}{group of checks where new check needs to be added}

\item{new_check}{new check defined by \code{new_check} function}
}
\description{
Add check into group
}
