% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{add_checks}
\alias{add_checks}
\title{Add list of checks into group}
\usage{
add_checks(group, new_checks)
}
\arguments{
\item{group}{group of checks where new check needs to be added}

\item{new_checks}{new check defined by \code{new_check} function}
}
\description{
Add list of checks into group
}
