% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporter.R
\name{check_definitions_to_data_frame}
\alias{check_definitions_to_data_frame}
\title{Convert group of checks into dataframe}
\usage{
check_definitions_to_data_frame(group)
}
\arguments{
\item{group}{group of checks to be converted}
}
\description{
Convert group of checks into dataframe
}
