% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runner.R
\name{check_result_info}
\alias{check_result_info}
\title{Select check result attribute from check run results as return vector of values}
\usage{
check_result_info(check_result, attribute)
}
\arguments{
\item{check_result}{list of check application outcomes (\code{apply_check})}

\item{attribute}{attribute name}
}
\description{
Select check result attribute from check run results as return vector of values
}
