% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporter.R
\name{check_results_to_data_frame}
\alias{check_results_to_data_frame}
\title{Convert check results into data frame}
\usage{
check_results_to_data_frame(check_results)
}
\arguments{
\item{check_results}{list of check results produced by \code{run_checks} 
function}
}
\description{
Convert check results into data frame
}
