% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-functions-column.R
\name{col_hasAllValues}
\alias{col_hasAllValues}
\title{All values from list present in column}
\usage{
col_hasAllValues(data, column, lov)
}
\arguments{
\item{data}{dataframe}

\item{column}{column name}

\item{lov}{list of values given as vector}
}
\description{
All values from list present in column
}
