% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-functions-column.R
\name{col_hasAnyValue}
\alias{col_hasAnyValue}
\title{Any of values from list present in column}
\usage{
col_hasAnyValue(data, column, lov)
}
\arguments{
\item{data}{dataframe}

\item{column}{column name}

\item{lov}{list of values given as vector}
}
\description{
Any of values from list present in column
}
