% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-functions-column.R
\name{col_hasApproxQuantile}
\alias{col_hasApproxQuantile}
\title{custom validation of a particular quantile of a column (approx.)}
\usage{
col_hasApproxQuantile(data, column, quantile, udf)
}
\description{
custom validation of a particular quantile of a column (approx.)
}
