% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-functions-column.R
\name{col_hasCompleteness}
\alias{col_hasCompleteness}
\title{Assess ratio of missing value with user-defined function}
\usage{
col_hasCompleteness(data, column, udf)
}
\arguments{
\item{data}{dataframe}

\item{column}{column name}

\item{udf}{user-defined function that takes ratio and returns logical value}
}
\description{
Assess ratio of missing value with user-defined function
}
