% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-functions-column.R
\name{col_hasConsistentType}
\alias{col_hasConsistentType}
\title{The largest fraction of values has the same type}
\usage{
col_hasConsistentType(data, column)
}
\description{
The largest fraction of values has the same type
}
