% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-functions-column.R
\name{col_hasCorrelation}
\alias{col_hasCorrelation}
\title{Has correlation with reference column}
\usage{
col_hasCorrelation(data, column, ref_column, udf)
}
\arguments{
\item{data}{dataframe to check}

\item{column}{column name}

\item{ref_column}{reference column name}

\item{udf}{user-defined function to apply to correlation value found}
}
\description{
Has correlation with reference column
}
