% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-functions-column.R
\name{col_hasCountDistinct}
\alias{col_hasCountDistinct}
\title{Has number of distinct non-null values}
\usage{
col_hasCountDistinct(data, column, udf)
}
\arguments{
\item{data}{dataframe}

\item{column}{column name}

\item{udf}{function that takes number and returns logical value}
}
\description{
Has number of distinct non-null values
}
