% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-functions-column.R
\name{col_hasDistinctness}
\alias{col_hasDistinctness}
\title{Assess distinctness of column values with user-defined function}
\usage{
col_hasDistinctness(data, column, udf)
}
\arguments{
\item{data}{dataframe}

\item{column}{column name}

\item{udf}{user-defined function that takes ratio and returns logical value}
}
\description{
Distinctness is defined by formula |V| / N where V is is total set of values that column takes, N is number of values in column.
}
