% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-functions-column.R
\name{col_hasEntropy}
\alias{col_hasEntropy}
\title{custom validation of a column’s entropy}
\usage{
col_hasEntropy(data, column, udf)
}
\description{
custom validation of a column’s entropy
}
