% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-functions-column.R
\name{col_hasHistogramValue}
\alias{col_hasHistogramValue}
\title{Has histogram value}
\usage{
col_hasHistogramValue(data, column, udf, ...)
}
\arguments{
\item{data}{dataframe to check}

\item{column}{column name to check}

\item{udf}{user-defined function to apply to histogram value found}

\item{...}{other parameters that are passed to \code{hist} function}
}
\description{
Function to build histogram object and examine it with user-defined function
}
