% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-functions-column.R
\name{col_hasMean}
\alias{col_hasMean}
\title{Has mean value}
\usage{
col_hasMean(data, column, udf)
}
\arguments{
\item{data}{dataframe to check}

\item{column}{column name to check}

\item{udf}{user-defined function to apply to mean value found}
}
\description{
Has mean value
}
