% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-functions-column.R
\name{col_hasMin}
\alias{col_hasMin}
\title{Has minimum value}
\usage{
col_hasMin(data, column, udf)
}
\arguments{
\item{data}{dataframe to check}

\item{column}{column name to check}

\item{udf}{user-defined function to apply to minimum value found}
}
\description{
Has minimum value
}
