% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-functions-column.R
\name{col_hasPredictability}
\alias{col_hasPredictability}
\title{user-defined validation of the predictability of a column}
\usage{
col_hasPredictability(data, column, ref_columns, udf)
}
\description{
user-defined validation of the predictability of a column
}
