% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-functions-column.R
\name{col_hasQuantile}
\alias{col_hasQuantile}
\title{custom validation of a particular quantile of a column}
\usage{
col_hasQuantile(data, column, probability, udf)
}
\arguments{
\item{data}{dataframe to check}

\item{column}{column name to check}

\item{probability}{probability for calculation of quantile}

\item{udf}{user-defined function to apply to quantile found}
}
\description{
Function is applicable only to numeric columns. It calculates value of
quantile of column values for a given probability. user defined function does
validation of quantile. Quantile is a value that corresponds to probability
of taking random value from sample that is less than quantile value
}
