% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-functions-column.R
\name{col_hasStandardDeviation}
\alias{col_hasStandardDeviation}
\title{custom validation of a column’s standard deviation}
\usage{
col_hasStandardDeviation(data, column, udf)
}
\description{
custom validation of a column’s standard deviation
}
