% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-functions-column.R
\name{col_hasType}
\alias{col_hasType}
\title{Column is of specific type}
\usage{
col_hasType(
  data,
  column,
  type = c("Date", "numeric", "integer", "logical", "character", "factor")
)
}
\arguments{
\item{data}{dataframe}

\item{column}{column name}

\item{type}{text name of the type - see \link{col_type}}
}
\description{
Column is of specific type
}
