% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-functions-column.R
\name{col_isComplete}
\alias{col_isComplete}
\title{No missing values in a column}
\usage{
col_isComplete(data, column)
}
\arguments{
\item{data}{dataframe}

\item{column}{column name}
}
\description{
No missing values in a column
}
