% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-functions-column.R
\name{col_isContainedIn}
\alias{col_isContainedIn}
\title{Value is contained in a list of predefined values}
\usage{
col_isContainedIn(data, column, alloved_values)
}
\arguments{
\item{data}{dataframe}

\item{column}{column name}

\item{alloved_values}{list of values given as vector}
}
\description{
Compatibility function that calls isInLOV
}
