% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-functions-column.R
\name{col_isInLOV}
\alias{col_isInLOV}
\title{Value is contained in a list of predefined values}
\usage{
col_isInLOV(data, column, lov)
}
\arguments{
\item{data}{dataframe}

\item{column}{column name}

\item{lov}{list of values given as vector}
}
\description{
Value is contained in a list of predefined values
}
