% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-functions-column.R
\name{col_isInRange}
\alias{col_isInRange}
\title{Value belongs to list of values}
\usage{
col_isInRange(data, column, range)
}
\arguments{
\item{data}{dataframe}

\item{column}{column name}

\item{lov}{list of values given as vector}
}
\description{
Compatibility function that calls isInLOV
}
