% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-functions-column.R
\name{col_isNonNegative}
\alias{col_isNonNegative}
\title{validation whether all values in a numeric column are non-negative}
\usage{
col_isNonNegative(data, column)
}
\description{
validation whether all values in a numeric column are non-negative
}
