% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-functions-column.R
\name{col_isUnique}
\alias{col_isUnique}
\title{No duplicates in a column}
\usage{
col_isUnique(data, column)
}
\arguments{
\item{data}{dataframe}

\item{column}{column name}
}
\description{
No duplicates in a column
}
