% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-functions-column.R
\name{col_satisfies}
\alias{col_satisfies}
\title{validation whether all rows matching 1st predicate also match 2nd predicate}
\usage{
col_satisfies(data, predicate)
}
\description{
validation whether all rows matching 1st predicate also match 2nd predicate
}
