% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporter.R
\name{convert_checks_to_df}
\alias{convert_checks_to_df}
\title{Convert group of checks into dataframe}
\usage{
convert_checks_to_df(group)
}
\arguments{
\item{group}{group of checks to be converted}
}
\description{
Convert group of checks into dataframe
}
