% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporter.R
\name{convert_run_results_to_df}
\alias{convert_run_results_to_df}
\title{Convert check results into data frame}
\usage{
convert_run_results_to_df(check_results)
}
\arguments{
\item{check_results}{list of check results produced by \code{run_checks}
function}
}
\description{
Convert check results into data frame
}
