% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporter.R
\name{convert_to_df}
\alias{convert_to_df}
\title{Convert to dataframe}
\usage{
convert_to_df(x)
}
\arguments{
\item{x}{data to be converted}
}
\description{
Can be applied to both groupcheck definitions or check results
}
