% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-functions.R
\name{hasAnyValue}
\alias{hasAnyValue}
\title{validation whether all values from list present in column}
\usage{
hasAnyValue(data, column, lov)
}
\description{
validation whether all values from list present in column
}
