% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-functions.R
\name{hasApproxQuantile}
\alias{hasApproxQuantile}
\title{custom validation of a particular quantile of a column (approx.)}
\usage{
hasApproxQuantile(data, column, quantile, udf)
}
\description{
custom validation of a particular quantile of a column (approx.)
}
