% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-functions.R
\name{hasColumn}
\alias{hasColumn}
\title{Function that checks whether column presents in the dataframe}
\usage{
hasColumn(data, column)
}
\arguments{
\item{data}{dataframe}

\item{column}{column name}
}
\description{
Function that checks whether column presents in the dataframe
}
