% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-functions.R
\name{hasColumnCount}
\alias{hasColumnCount}
\title{Function that calculates number of columns in dataframe and applies UDF (user
defined function) to it}
\usage{
hasColumnCount(data, udf)
}
\arguments{
\item{data}{dataframe}

\item{udf}{user-defined funcion that takes number and returns logical value}
}
\description{
Function that calculates number of columns in dataframe and applies UDF (user
defined function) to it
}
