% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-functions.R
\name{hasColumns}
\alias{hasColumns}
\title{Function that checks whether columns presents in the dataframe}
\usage{
hasColumns(data, columns)
}
\arguments{
\item{data}{dataframe}

\item{columns}{vector of column names}
}
\description{
Function returns TRUE if all columns present in the dataframe
}
