% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-functions.R
\name{hasConsistentType}
\alias{hasConsistentType}
\title{The largest fraction of values has the same type}
\usage{
hasConsistentType(data, column)
}
\description{
The largest fraction of values has the same type
}
