% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-functions.R
\name{hasCountDistinct}
\alias{hasCountDistinct}
\title{custom validation of the number of distinct non-null values in a column}
\usage{
hasCountDistinct(data, column, udf)
}
\description{
custom validation of the number of distinct non-null values in a column
}
