% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-functions.R
\name{hasMax}
\alias{hasMax}
\title{custom validation of a column’s maximum value}
\usage{
hasMax(data, column, udf)
}
\arguments{
\item{data}{dataframe to check}

\item{column}{column name to check}

\item{udf}{user-defined function to apply to maximum value found}
}
\description{
custom validation of a column’s maximum value
}
