% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-functions.R
\name{hasPredictability}
\alias{hasPredictability}
\title{user-defined validation of the predictability of a column}
\usage{
hasPredictability(data, column, ref_columns, udf)
}
\description{
user-defined validation of the predictability of a column
}
