% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-functions.R
\name{hasRowCount}
\alias{hasRowCount}
\title{Function that calculates number of rows in dataframe and applies UDF (user
defined function) to it}
\usage{
hasRowCount(data, udf)
}
\arguments{
\item{data}{dataframe}

\item{udf}{user-defined funcion that takes number and returns logical value}
}
\description{
Function that calculates number of rows in dataframe and applies UDF (user
defined function) to it
}
