% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-functions.R
\name{hasStandardDeviation}
\alias{hasStandardDeviation}
\title{custom validation of a column’s standard deviation}
\usage{
hasStandardDeviation(data, column, udf)
}
\description{
custom validation of a column’s standard deviation
}
