% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-functions.R
\name{hasTypeConsistency}
\alias{hasTypeConsistency}
\title{custom validation of the maximum fraction of values of the same data type}
\usage{
hasTypeConsistency(data, column, udf)
}
\description{
custom validation of the maximum fraction of values of the same data type
}
