% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-functions.R
\name{hasUniqueKey}
\alias{hasUniqueKey}
\title{Function that checks whether combination of columns contains unique value}
\usage{
hasUniqueKey(data, columns)
}
\arguments{
\item{data}{dataframe}

\item{columns}{vector of column names that must contain unique value}
}
\description{
Function that checks whether combination of columns contains unique value
}
