% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-functions.R
\name{hasValue}
\alias{hasValue}
\title{validation whether column has values that satisfy predicate and uder defined function}
\usage{
hasValue(data, column, predicate)
}
\description{
validation whether column has values that satisfy predicate and uder defined function
}
