% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-functions.R
\name{isComplete}
\alias{isComplete}
\title{No missing values in a column}
\usage{
isComplete(data, column)
}
\arguments{
\item{data}{dataframe}

\item{column}{column name}
}
\description{
No missing values in a column
}
