% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-functions.R
\name{isGreaterThan}
\alias{isGreaterThan}
\title{validation whether values in the 1s column are greater than in the 2nd column}
\usage{
isGreaterThan(data, column, ref_column)
}
\description{
validation whether values in the 1s column are greater than in the 2nd column
}
