% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-functions.R
\name{isInLOV}
\alias{isInLOV}
\title{Value belongs to list of values}
\usage{
isInLOV(data, column, lov)
}
\arguments{
\item{data}{dataframe}

\item{column}{column name}

\item{lov}{list of values given as vector}
}
\description{
Value belongs to list of values
}
