% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-functions.R
\name{isNonNegative}
\alias{isNonNegative}
\title{validation whether all values in a numeric column are non-negative}
\usage{
isNonNegative(data, column)
}
\description{
validation whether all values in a numeric column are non-negative
}
