% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-functions.R
\name{isNotLessThan}
\alias{isNotLessThan}
\title{validation whether values in the 1s column are not less than in the 2nd column}
\usage{
isNotLessThan(data, column, ref_column)
}
\description{
validation whether values in the 1s column are not less than in the 2nd column
}
