% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-functions.R
\name{isUnique}
\alias{isUnique}
\title{No duplicates in a column}
\usage{
isUnique(data, column)
}
\arguments{
\item{data}{dataframe}

\item{column}{column name}
}
\description{
No duplicates in a column
}
