% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapter.R
\name{load_group}
\alias{load_group}
\title{Load group of checks from RDS file}
\usage{
load_group(file_name)
}
\arguments{
\item{file_name}{source filename that contains group of checks}
}
\description{
Load group of checks from RDS file
}
