% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{new_check}
\alias{new_check}
\title{Create new check}
\usage{
new_check(description, severity, function_name, ...)
}
\arguments{
\item{description}{check description expressed in understandable way.
it should give good idea of a purpose of the check. Description should define
condition that data must meet. Like "Country must be always filled".
Good practice is to a) start with name of subject of check b) give clear and
concise criteria c) specify when it is applicable (if not always)}

\item{severity}{check severity, can be chosen from `severity` list attached
to package}

\item{function_name}{validation function name - text or function (without parameters)}

\item{...}{other parameters that will be passed to validation function like 
\code{column_name}, \code{udf}, etc.}
}
\description{
Function takes parameters and returns data structure that represents check
and can de included into group and then executed
}
