% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{new_checks_for_columns}
\alias{new_checks_for_columns}
\title{Add same check for different columns}
\usage{
new_checks_for_columns(columns, description, severity, function_name, ...)
}
\arguments{
\item{columns}{vector of columns}

\item{description}{check description}

\item{severity}{check severity, can be chosen from `severity` list attached
to package}

\item{function_name}{validation function name - text or function (without parameters)}

\item{...}{other parameters that will be passed to validation function like 
\code{udf}, etc.}
}
\description{
Similar to \link{new_check} but allows to specify check to many columns in 1 go
}
