% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-utils.R
\name{ratio}
\alias{ratio}
\title{Ratio of success based on vector of logical values}
\usage{
ratio(logical_vector)
}
\arguments{
\item{logical_vector}{vector of logical values}
}
\description{
Ratio of success based on vector of logical values
}
