% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{run_check}
\alias{run_check}
\title{Execute check one check on data}
\usage{
run_check(data, check)
}
\arguments{
\item{data}{dataset to be checked}

\item{check}{check to apply - use \code{new_check} function to create a check}
}
\description{
Function executes check on given data and creates data structure (list) with
check summary
}
