% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runner.R
\name{run_checks}
\alias{run_checks}
\title{Apply checks to dataset and return dataframe}
\usage{
run_checks(data, checks, condition = NULL)
}
\arguments{
\item{data}{dataframe to check}

\item{checks}{group of checks created by new_group() %>% add_check() ...}

\item{condition}{function that takes check_result
and returns TRUE / FALSE where FALSE halts execution.}
}
\description{
Apply group of checks to dataset. Function returns check results. In case when
\code{stop_condition} is specified function stops execution when
}
