% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runner.R
\name{run_checks_and_proceed}
\alias{run_checks_and_proceed}
\title{Apply all checks to dataset and if there are no errors}
\usage{
run_checks_and_proceed(data, checks, condition)
}
\arguments{
\item{data}{dataset that needs to be checked}

\item{checks}{group of checks that needs to be applied to dataset}

\item{condition}{user defined function that needs to take check results and
return either TRUE or FALSE}
}
\description{
This convinience function combines two: \code{run_checks} and
\code{analyze_run}. Check resuts are given to analyze function. If analyze
finds violations that exceed allowed threshold then process stops. Otherwise
function will return initial dataset and processing can be continued.
}
