% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-functions.R
\name{satisfies}
\alias{satisfies}
\title{validation whether all rows matching 1st predicate also match 2nd predicate}
\usage{
satisfies(data, predicate)
}
\description{
validation whether all rows matching 1st predicate also match 2nd predicate
}
