% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-functions.R
\name{satisfiesIf}
\alias{satisfiesIf}
\title{validation whether all rows matching 1st predicate also match 2nd predicate}
\usage{
satisfiesIf(data, predicate_1, predicate_2)
}
\description{
validation whether all rows matching 1st predicate also match 2nd predicate
}
