% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapter.R
\name{save_check}
\alias{save_check}
\title{Save check into RDS file}
\usage{
save_check(check, file_name)
}
\arguments{
\item{check}{check to be exported}

\item{file_name}{destination filename}
}
\description{
Save check into RDS file
}
