% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapter.R
\name{save_group}
\alias{save_group}
\title{Save group of checks into RDS file}
\usage{
save_group(group, file_name)
}
\arguments{
\item{group}{group of checks to be exported}

\item{file_name}{destination filename}
}
\description{
Save group of checks into RDS file
}
