% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{severity_above_threshold}
\alias{severity_above_threshold}
\title{Function that creates severity check function.}
\usage{
severity_above_threshold(severity)
}
\arguments{
\item{severity}{- severity text label (INFO, WARNING, ERROR)}
}
\description{
Function that creates severity check function.
}
