% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{severity_rank}
\alias{severity_rank}
\title{Translate severity label into rank}
\usage{
severity_rank(severity)
}
\arguments{
\item{severity}{severity value (or vector of values) from \code{severity}
list}
}
\description{
Function takes vector of severity labels and translates them into number
according to their position within \code{severity} list. Severities in list
start from lowest to highest. So first mentioned severity has rank = 1 and
every next severity has rank incremented by one
}
