% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runner.R
\name{severity_under_threshold}
\alias{severity_under_threshold}
\title{Create severity check function.}
\usage{
severity_under_threshold(severity)
}
\arguments{
\item{severity}{- severity text label (INFO, WARNING, ERROR)}
}
\description{
Function will check and return TRUE when severity is under threshold for
checks with \code{check_sucess} status = \code{FALSE}
}
