% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-utils.R
\name{stop_if_miss_columns}
\alias{stop_if_miss_columns}
\title{Raise error in case when dataframe doesn't contaain columns}
\usage{
stop_if_miss_columns(data, columns)
}
\arguments{
\item{data}{dataframe}

\item{columns}{vector of column names}
}
\description{
Raise error in case when dataframe doesn't contaain columns
}
