% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-utils.R
\name{stop_if_not_implemented}
\alias{stop_if_not_implemented}
\title{Function that just stops execution and generates error message}
\usage{
stop_if_not_implemented(name, not_implemented)
}
\arguments{
\item{name}{- name of the function which is not implemented}

\item{not_implemented}{- logical value TRUE / FALSE}
}
\description{
Function that just stops execution and generates error message
}
