% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-functions-table.R
\name{tab_hasColumn}
\alias{tab_hasColumn}
\title{Function that checks whether column presents in the dataframe}
\usage{
tab_hasColumn(data, column)
}
\arguments{
\item{data}{dataframe}

\item{column}{column name}
}
\description{
Function that checks whether column presents in the dataframe
}
