% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-functions-table.R
\name{tab_hasColumnOfType}
\alias{tab_hasColumnOfType}
\title{Function that checks whether dataframe column is of specific type}
\usage{
tab_hasColumnOfType(data, column, type = c("Date", "numeric", "integer",
  "logical", "character", "factor"))
}
\arguments{
\item{data}{dataframe}

\item{column}{column name}

\item{type}{text name of the type - see \link{col_type}}
}
\description{
Function that checks whether dataframe column is of specific type
}
