% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-helpers.R
\name{udf_between}
\alias{udf_between}
\title{Between function}
\usage{
udf_between(x1, x2, include = "both")
}
\arguments{
\item{x1}{- left border}

\item{x2}{- right border}

\item{include}{- inclusion type (0 / none, 1 / left, 2 / right, 3 / both) 
can be specified with digit, words, first letter of words}
}
\description{
Function returns another function that checks if given value is in between
\code{x1} and \code{x2}
}
