% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-helpers.R
\name{udf_eq}
\alias{udf_eq}
\title{Equal to function}
\usage{
udf_eq(x)
}
\arguments{
\item{x}{- value to compare with.}
}
\description{
Function returns another function that checks if given value
is equal to \code{x}
}
