% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-helpers.R
\name{udf_ge}
\alias{udf_ge}
\title{Greater than or equal to function}
\usage{
udf_ge(x)
}
\arguments{
\item{x}{- value to compare with.}
}
\description{
Function returns another function that checks if given value is greater than
or equal \code{x}
}
