% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-helpers.R
\name{udf_gt}
\alias{udf_gt}
\title{Greater than function}
\usage{
udf_gt(x)
}
\arguments{
\item{x}{- value to compare with.}
}
\description{
Function returns another function that checks if given value is greater than
to \code{x}
}
\examples{
hasMean(data = data.frame(col1 = [0:10]), udf = udf_gt(0))
}
