% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-helpers.R
\name{udf_le}
\alias{udf_le}
\title{Less than or equal to function}
\usage{
udf_le(x)
}
\arguments{
\item{x}{- value to compare with.}
}
\description{
Function returns another function that checks if given value is less than 
or equal to \code{x}
}
