% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-helpers.R
\name{udf_lt}
\alias{udf_lt}
\title{Less than function}
\usage{
udf_lt(x)
}
\arguments{
\item{x}{- value to compare with.}
}
\description{
Function returns another function that checks if given value is less than
\code{x}
}
