% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runner.R
\name{analyze_run}
\alias{analyze_run}
\title{Analyze checks results and stop if data quality is insufficient}
\usage{
analyze_run(check_result, condition)
}
\arguments{
\item{check_result}{check results that need to be analyzed}

\item{condition}{function that takes check_result and returns TRUE / FALSE
where TRUE means success and proceeds execution and FALSE halts execution.}
}
\description{
Data quality reported in check resutls must be sufficient otherwise function
will raise error and stop program execution. Validation condition is given as
anonymous function into \code{condition}
}
