% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-functions-column.R
\name{col_hasAnyValue}
\alias{col_hasAnyValue}
\title{validation whether all values from list present in column}
\usage{
col_hasAnyValue(data, column, lov)
}
\description{
validation whether all values from list present in column
}
