% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-functions-column.R
\name{col_hasCountDistinct}
\alias{col_hasCountDistinct}
\title{custom validation of the number of distinct non-null values in a column}
\usage{
col_hasCountDistinct(data, column, udf)
}
\description{
custom validation of the number of distinct non-null values in a column
}
