% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-functions-column.R
\name{col_hasMin}
\alias{col_hasMin}
\title{custom validation of a column’s minimum value}
\usage{
col_hasMin(data, column, udf)
}
\description{
custom validation of a column’s minimum value
}
