% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-functions-column.R
\name{col_hasTypeConsistency}
\alias{col_hasTypeConsistency}
\title{custom validation of the maximum fraction of values of the same data type}
\usage{
col_hasTypeConsistency(data, column, udf)
}
\description{
custom validation of the maximum fraction of values of the same data type
}
