% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-functions-column.R
\name{col_isLessThan}
\alias{col_isLessThan}
\title{validation whether values in the 1s column are less than in the 2nd column}
\usage{
col_isLessThan(data, column, ref_column, value)
}
\description{
validation whether values in the 1s column are less than in the 2nd column
}
