% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-functions-column.R
\name{col_satisfies}
\alias{col_satisfies}
\title{Data rows match the predicate}
\usage{
col_satisfies(data, predicate)
}
\arguments{
\item{data}{dataframe}

\item{predicate}{check function that takes \code{data} as a parameter and
returns TRUE if records satisy predicate of FALSE if not}
}
\description{
Data rows match the predicate
}
